Create additional facts for postgres configuration
These facts have been added to support modification of the default
mode of operation for management of postgres in the situation where
another management tool requires to manage part or whole of postgres.
This is primarily designed to enable Patroni and to avoid conflicts
and configuration management flap between successive runs with TPA
and Patroni. Wherever possible the configuration source of truth remains
with TPA and users should still be able to use the standard documented
interfaces via TPA's `config.yml`. The new fact flags for management
of postgres components are not intended for direct use by a user and
are enabled by either activating features in `config.yml` or configure
or by automated gathering of the state of the cluster.

Provides the following boolean facts:

    postgres_config_split: true
    postgres_create_hba_file: true
    postgres_create_ident_file: true
    postgres_service_managed: true
    postgres_service_enabled: true

A fact for the `postgresql.conf` cluster_name config setting.
A switch for managed postgres config include files and systemd service.
A fact for enabling creation of `pg_hba.conf` and `pg_ident.conf`.

Postgres config include files are created only when the fact
`postgres_config_split` is true, the default. If it is not true,
which is altered by the patroni role, then the include files can
be removed by calling `cleanup.yml` tasks. This is not called by
default.

Provides the following configuration object facts:

    postgres_pg_hba_config
    postgres_pg_ident_config
    postgres_config

To drive configuration of postgres in alternative tooling new facts
are created which expose the tunable aspects of postgres. These facts
make copies of data held in include template files. As the only user of
these facts is currently Patroni the template files themselves do not
use these facts. The facts live in the postgres role and not in the
patroni role to allow access the role defaults and ensure the ordering
of fact setting is in line with related setup processes that inform the
configuration of some values, for example extension setup. Some data is
not available until files have been created either by dependent parts
of the postgres/config role or installation steps.

Without patroni enabled, these steps will not alter normal or existing
operation of TPA. The facts are still created, but they do not influence
how the cluster is managed.

References: TPA-398, TPA-400, TPA-417