Commit changes made by relink to git
We're running relink automatically on upgrade to replace BDR-Always-ON's
upgrade.yml with PGD-Always-ON's upgrade.yml, which causes provision to
bail out because the change to commands has not been committed. This is
arguably something that reconfigure should be doing when changing the
architecture, but we choose the easier fix for now.