Wait for write leader elections to complete before starting pgd-proxy
We can't know for sure in advance if pgd-proxy will start correctly, but
since we have to stop harp-proxy before starting it, we might as well do
whatever we can to help it along. Waiting for all subgroups to have a
write leader, as we do here, means pgd-proxy will be ready to route
traffic sooner after it's started.

Also, if for some reason BDR is unable to elect a write leader for the
newly-creating routing groups, we now fail without stopping harp-proxy.