Misc documentation updates
* Consolidate distribution support info into distributions.md

The distinction drawn in the docs between host system and target system
support no longer seems to be required.
Consolidating this information helps clarity and reduces the number of
places to update when new distros are supported.

* Refactor install.md codeblocks for EDB style

The EDB docs site generally provides code snippets that can be directly
executed. This means no prompt characters or code for different distros in the same
block. I have changed install.md to this style experimentally.
---------

Co-authored-by: Dj Walker-Morgan <126472455+djw-m@users.noreply.github.com>
Co-authored-by: Haroon <muhammad.haroon@enterprisedb.com>