Require at least BDR v4.3 as the source for major version upgrades
We know the upgrade process requires at least PGD 5.2 (because we need
its fixes for bdr.switch_node_group() to upgrade clusters that include
witness/standby/etc. nodes).

We now require that the cluster must be running at least 4.3 to start
with (4.3 was released in February, and 4.3.1 was released in May),
because older versions don't have an extension upgrade path to 5.2.

Reported by Jonathan from failing CI tests on a Debian 11 cluster with
EPAS14. The cluster started with 4.2.2, because apparently the debian:11
packages for 4.3 were not correctly published to products/bdr4/release.
In this scenario, restarting the Postgres server after the upgrade would
appear to work, but the upgrade would fail shortly afterwards:

    TASK [postgres/facts : Collect facts about the Postgres cluster] …
    task path: /home/ams/work/edb/tpaexec/roles/postgres/facts/tasks/main.yml:8
    An exception occurred during task execution. To see the full traceback, use -vvv. The error was:                                       ^
    fatal: [unseats]: FAILED! => {"changed": false, "error": "relation \"bdr.node_config\" does not exist\nLINE 2:                     LEFT JOIN bdr.node_config nc USING (node...\n                                      ^\n"}

The reason the v5-specific bdr.node_config catalog doesn't exist despite
running the v5.2 binaries is because the extension upgrade failed:

    Jul 20 16:03:27 unseats postgres[3796]: [13782-1] 2023-07-20 16:03:27 UTC [enterprisedb@/pglogical manager 16926/bdrdb:3796]: [2] ERROR:  extension "bdr" has no update path from version "4.2.2" to version "5.2.0-dev"

This is an unpleasant thing to discover *after* stopping Postgres, so
it's worth trying to avoid it.