Ignore "Depsolve" errors in postgres/pkg:check-update
We want to warn the user about missing packages early enough for them to
do something about it, not confuse them with errors like the following
one when the required packages are in fact present:

    fatal: [ushers]: FAILED! => {"changed": false, "msg": "Package installation check failed: Depsolve Error occured: \n Problem: problem with installed package edb-bdr4-epas14-4:4.3.1-5.el8.x86_64\n  - package edb-bdr5-epas14-4:5.2.0-0.0snapshot5613502528.1002.1.bdc811d.0.el8.x86_64 conflicts with edb-bdr4-epas14 provided by edb-bdr4-epas14-4:4.3.1-5.el8.x86_64\n  - conflicting requests"}

The conflict is nothing unexpected, and isn't relevant to the upgrade
because we're going to uninstall the bdr4 package before installing the
bdr5 ones.

Rerun the upgrade with `-e check_packages_before_upgrade=no` to skip the
check, if the error message is something other than a "Depsolve Error"
that's still ultimately not a problem.