Add default subscriber-only group to bdr_node_groups for subscriber-only instances
Although BDR4 clusters did not (to our knowledge) feature any explicit
use of bdr_node_groups in config.yml (even though this was supported,
after a fashion), we did add a default subscriber-only group if there
were subscriber-only instances that didn't declare a bdr_child_group
explicitly.

We do have customers that depended on this behaviour. The group we added
would exist in the database, of course; here we merely make it explicit
in the configuration, and make sure to set the right bdr_child_group, so
that they don't try to switch to one of the location-specific subgroups
that we create here (that would change what they subscribe to, but we
don't want to change their behaviour).