Invoke postgres/user during the node-by-node upgrade
When switching from the old 2ndQPostgres/PGE packages to the current
edbpge ones, we have to change unix_socket_directories. The value is
also written into ~postgres/.bashrc by postgres/user, so we invoke it
during the upgrade. (It was originally included in the TPA-387 branch,
but the commit message didn't explain why it had been added, so I took
it out to see if anything broke.)