Avoid using list-client-packages to install psycopg2 for testing
We might be testing harp-proxy as part of an upgrade on a cluster
that's currently running 2ndQPostgres/PGE, but where config.yml already
sets `postgresql_flavour: edbpge`. At this stage of the upgrade, we have
not yet set up the repositories, so we can't expect installation of the
edbpge client packages to work.

We really need only psycopg2 for the test (and let it install whatever
libpq it can find, per the declared package dependencies), so in order
to avoid having to detect what flavour we're running and figure out how
to override the flavour when including list-client-packages, we just
drop the client packages from the request.

References: TPA-518