Keep pemagent barman config on redeploy
When we set up pemagent on a barman server, we write a config file based
on a template and then use pemworker to add the barman server to pem,
which appends to the config file. If we later run deploy again, we will
write the config file again, but we can't rerun pemworker, so we end up
with a config file missing the barman section.

We overcome this problem by looking for the barman section early on, and
then if we have found it, writing it as part of the templating.