TPA-523: Fix replica of SO
Fix the assertion that checked replica's roles
instead of upstream's roles.

the Objective of this check was to block creation of bdr node replica
but still allow replica of subscriber-only node.
the isssue was that the test was applied to the replica node instead of
upstream node that will be replicated.

in order to have a replica for subscriber node the replica should
only require `replica` role and `bdr`, and upstream node being
replicated needs to be `bdr`, `subscriber-only`
as shown here:
```
- Name: kipper
  location: a
  node: 5
  role:
  - bdr
  - subscriber-only
- Name: replicator
  location: b
  node: 14
  role:
  - replica
  - bdr
  upstream: kipper
```