Document `cleanup` tag for `download-packages` command
If `download-packages` command fails during the download process, it could
leave behind a running container used for downloading the packages for
debugging purposes. If needed, we can get rid of the container by running
`download-packages` command with `--tags cleanup`. The functionality has
existed for as long as the `download-packages` command itself, it just
wasn't documented.

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>