Construct `base_image` from the image name plus the tag in a fact
When using the intermediate base docker image option use the name of the
image plus the tag to construct the `BASE_IMAGE` arg to the final
image builder.

It appears that the digest form `sha256:xxxx` only works when an image
has a digest, which is created when an image is pushed to a v2 registry.
An image that is built locally will have no digest and specifying one
will force the docker daemon to attempt to retrieve the image from a
registry.

References: TPA-536