Make java jdk 11 the default for EFM for supported operating systems
Added exceptions for Debian-9 and Ubuntu-18, to retain backwards
compatibility.

This changes the previous arrangement, where an OS release version
specific file exists with the latest package names and the default falls
back to an older package names. With the new layout, the default for
unspecified OS versions is the latest and older OS releases have
specific files mentioning to the older jdk version package. Assuming the
package names remaining the same, as they have done for the last couple
of releases of Debian and RHEL, we are less likely need to update these
files in the future, thus reducing the maintenance overhead.

References: TPA-548