Require Python 3.9+ when running bin/tpaexec
We maintain backwards compatibility with Python 3.6 so far, because both
RHEL7 and RHEL8 have that version by default. But with the cryptography
module removing support for that version, and several pending security
alerts, we want to begin moving away from 3.6 ourselves.

This commit sets the stage for automatically discovering another Python
interpreter (e.g., from edb-python, if available) if the system's
default one is too old. Otherwise, if you initialise your venv with a
different interpreter, or another python is in your PATH, it should just
work sensibly without further ado.

References: TPA-562