Set a default mininmum `wal_keep_size` of 16MB for patroni
The old default of `0` is considered an invalid setting in Patroni,
it requires a minimum of 16MB. Setting anything less results in a
warning message in the logs and on the command line when using
`patronictl`.

This only changes the config stored in `postgres_restart_config` fact
and not that used for standard postgres file based config.

References: TPA-560