Add support for BDR3.7 to PGD5 upgrade
Add reconfigure support to generate the PDG5 cluster config.yml from
BDR-Always-ON cluster with version 3.7. Reconfigure from bdr3 to pgd5
includes:
- removal of pglogical from `extra_postgres_extensions`
- accepting bdr_version 3 as supported version for `reconfigure`

Also split the existing upgrade to determine the type of upgrade we are
running and then fork to the relevant upgrade playbook depending on the
type determined e.g. major 3.7 to 5 upgrade or a minor PGD5 upgrade.
Keeps generic initial steps into main upgrade.yml for PGD-Always-ON.

Moves the existing upgrade process for 4 to 5 and minor upgrade for 5 to
upgrade_legacy.yml. Later, this legacy_upgrade.yml is to be split into
minor and major upgrades too. See TPA-566 and TPA-567 for details.

Modify package removal selection for bdr version based on starting
version instead of target version. Also ensure we correctly detect the
postgres packages for bdr version 3 to pgd 5 upgrade when it involves a
flavour change from legacy pgextended to edbpge. Also replace the
upgrade_type to upgrade_from which simplifies the implementation logic.

Changes to ensure that legacy pgextended to edbpge flavour switch is
correctly handled during the upgrade. For example, the reconfigured
config.yml will have moved to `edbpge` while the old packages that we
want to remove from the target nodes during the upgrade are legacy
pgextended packages which use different names; this requires
constructing the list of legacy pgextended and corresponding bdr3
package names so upgrade can correctly remove them.

We used to only remove bdr related packages and let pgextended binaries
stay. Change that to ensures we list pgextended packages and remove them
before installing edbpge.

Ensure PEM server postgres is also updated if there is a PEM instance in
the cluster.

Backport upgrade improvements introduced during BDR4 to PGD5 upgrade
implementation; check that the BDR major version currently in use
matches the one in config.yml; the logic to do this is backported from
the similar check in the PGD-Always-ON version of upgrade.yml. Includes
Fence/unfence process using harpctl introduced in BDR4 to PGD5 upgrade;
and component specific upgrade.yml tasks from the individual components.

Since the upgrade process uses pgrep to determine running processes,
check that pgrep is installed at the beginning of the process instead of
attempting to use it when needed and possibly failing with the upgrade
half-done.

Relevant documentation updates.

References: TPA-445, TPA-565, TPA-568, TPA-570, TPA-508, TPA-552

Co-authored-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>
Co-authored-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>
Co-authored-by: Richard Poole <richard.poole@enterprisedb.com>
Signed-off-by: Jonathan Renon <98327055+JonathanRenon-EDB@users.noreply.github.com>