TPA-565: Initial split of upgrade.yml
Split the existing upgrade to prepare work on upgrade 3 to 5
keep generic initial steps into main upgrade.yml for PGD-Always-ON.
move the actual existing upgrade process for 4 to and minor upgrade in
a new file upgrade_legacy.yml

This legacy_upgrade.yml will be split into minor and major upgrade later.
see TPA-566 and TPA-567.

The goal is to keep a working upgrade while the split is in progress.