Accept `repmgr` as failover_manager in BDR when a replica is required.
Some customers might need to stay with repmgr for different reasons.
Without this TPA was setting the failover_manager as `none` when someone
wanted to go down this path. One of the reasons is if inside a bdr cluster
a node is required to work as a replica of a subscriber-only host, in
that case only, tpa will allow repmgr to be installed and configured.

References: TPA-578