Make sar run on all deployed machines, not only on Barman
TPA had a `sys/sysstat` role which was in charge of configuring `sar`
to run every minute, and enabling it.

However, the implementation would rely on `cron` service to be running,
which was only started in nodes with the `barman` role.

Through this commit we amend that behavior in such a way `sar` will
run on all nodes, not only on nodes that have the `barman` role.

We took advantage of that change to also modify how `sysstat` is
configured in the servers: instead of using cronjobs, we will now
use `systemd` timers. This is the approach that has been adopted for
`sysstat` package on newer distros like Debian 11 and RHEL 9, so
we changed TPA code to also do that.

In summary, this is what TPA will do on all supported distros:

* Create a `sysstat.service` unit, which is used as an entry point
  for booting `sa1` and installing `sysstat-collect.timer` and
  `sysstat-summary.timer` units
* Create `sysstat-collect.service` unit, which takes care of
  executing one round of 1 minute of `sa1`
* Create `sysstat-collect.timer` unit, which takes care of triggering
  the corresponding service every minute
* Create `sysstat-summary.service` unit, which performs an aggregation
  of metrics through `sa2`
* Create `sysstat-summary.timer` unit, which takes care of triggering
  the corresponding service every day at 7 o'clock

A few things needed to be taken into consideration:

* Debian/Ubuntu have the `sar` binaries under `/usr/lib/sysstat`
* RHEL/SLES have the `sar` binaries under `/usr/lib64/sa`
* `sa1` binary is named `debian-sa1` on Debian 10 and Ubuntu Focal
* When `sysstat` package is installed on Debian 10, Ubuntu Focal, or
  RHEL 7, the `sysstat.service` unit already comes enabled by default,
  so we need to disable it before enabling and starting it through
  `systemd` Ansible module, because attempting to enable an unit
  that is already enabled won't work as expected

References: TPA-595.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>