Do not toggle `sysstat` service on every `tpaexec deploy`
As reported by Abhijit during the PR review, previous to this commit
every execution of `tpaexec deploy` would cause `sysstat` service to
be disabled and re-enabled -- when deploying Debian 10, Ubuntu Focal
or RedHat 7.

This commit fixes that issue, so the disable/re-enable approach is
only used on the very first execution of `tpaexec deploy` for the
cluster. We only use that workaround if the custom `sysstat-collect.timer`
service has never been installed by TPA.

References: TPA-595.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>