Backport upgrade improvements to BDR-Always-ON
When upgrading a BDR-Always-ON cluster, check that the BDR major version
currently in use matches the one in config.yml; the logic to do this is
backported from the similar check in the PGD-Always-ON version of
upgrade.yml.

Fence/unfence the node being upgraded using harpctl.

Use the separate upgrade.yml tasks from the individual components as
appropriate.