TPA-614: Fix camo enabled cluster upgrade
The fix consists in starting postgres before runing postgres/config role
in order to gather updated facts through postgres/facts and postgres/bdr
version.yml to ensure TPA uses the correct values to configure postgres,
bdr, CAMO and all related pieces.
once facts are updated, we turn off postgres and continue with
postgres/config.

for 3to5 upgrade there are a couple entries that need to be removed from
3334-camo.conf and 8888-shared_preload_libraries.conf

for pgextended to edbpge upgrade we also need to fix the binary path used
by the postgres unit file and run daemon-reload.

This fix should also be used for 4to5 even if there is no breaking conf
set in this case for now.