WIP: POC for new way of deploying Patroni clusters
This is a POC with changes to change the way Patroni clusters are
deployed. So far TPA would do the following:

* TPA: bootstrap the primary
* Patroni: take over management of the primary
* Patroni: bootstrap replicas

With the changes it will now do the following:

* TPA: bootstrap primary and replicas
* Patroni: take over management of the cluster

That's better aligned with how other HA tools are deployed by TPA.

More commits should likely come later, at least containing changes
in the code comments which are still describing the previous
behavior.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>