Support choosing Ansible version
At setup-time, you can now choose between the community and 2ndQuadrant
versions of Ansible 2.9, or community Ansible 8 or 9. The default
Ansible is version 8, because this is the most recent version compatible
with Python 3.9 .

To enable TPA to run under recent ansible versions, there are various
minor changes to module invocations and to TPA's custom plugins. These
remain compatible with Ansible 2.9 .

Integration tests now allow selection of Ansible version for the
single-test and full workflows.