Add GH workflows and integration tests
In a recent shuffle to make TPA public, integration tests were moved out
of the main TPA source repository to restrict access. However, now that
we want to move to a public/private repo pair for TPA development for
various reasons, it makes sense to add workflows and integration tests
to the private repo to make development testing easier.

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>