Refactor extension configuration interface
The change teaches TPA to automatically include shared preload library
entries and package names for a select set of internally recongised
extensions when they are specified in the config.yml.

It also means that users only need to specify their desired extensions
in the `extra_postgres_extensions` list or the `extensions` list of a
database specified in `postgres_databases`.

The respective packages and any required `shared_preload_library` config
entries are contained in a dictionary. TPA handles installing the
packages, addition of any preloads and the corresponding `CREATE
EXTENSION $ext_name`.

References: TPA-388, TPA-293, TPA-307, TPA-336 and TPA-645