Store the ansible-vault password in the system keyring by default
We used to write the vault password to vault/vault_pass.txt, and we'll
continue to work with existing clusters where that is the case. For new
clusters, we'll try to use the keyring module to store the password into
any available system keyring (e.g., gnome-keyring).

Use `tpaexec configure … --keyring-backend <name>` to select a backend,
but the only available choices are "system" (the default) and "legacy"
(which is to use vault/vault_pass.txt).

vault/vault_pass.txt method stores credentials inside the cluster
directory and therefore there is no chance of conflict when multiple
clusters using the same name are provisioned on a single tpa host, it's
a challenge though when using a shared service. The change introduces an
additional configuration setting named `vault_name` for the config.yml
to go with the `keyring_backend`. For new clusters configured using
`tpaexec configure` command, `vault_name` will be set to a UUID to make
sure the combination of `cluster_name` and `vault_name` gives us a
unique combination when `system` (default) is used as the
keyring_backend. `vault_name` can be set to any arbitrary value to make
sure the combination of `cluster_name` and `vault_name` is unique. It
does not have to comply with UUID format.

Also update tpaexec-configure.md to document `keyring_backend` and
`vault_name` settings.

Adds `show-vault` command; it displays the vault password for both
`legacy` and `system` backends.

With contributions from Abhijit and Haroon.

References: TPA-85