Add `relnotes.yml` workflow to automate relnote compilation
This commit installs the `relnotes.yml` workflow and the related
scripts -- taken from github.com/EnterpriseDB/release_notes repo.

The idea of the workflow is to automate the generation of release
notes when performing releases of TPA.

This is an overview of the process when using this automation:

* For each ticket to be worked on:
  * Developer creates a dev branch
  * Developer writes the code
  * Developer writes the release notes of that ticket in YAML format
     under `release_notes` folder in the TPA repo
  * Developer opens the PR both with the code changes and with the
    corresponding release notes
  * Reviewer reviews and test the code and notes. Once the PR is
    merged, the `main` branch will contain the release notes of the
    ticket
* When it is time for releasing:
  * Release manager triggers the execution of the `relnotes.yml`
    workflow. That will create a PR against the `main` branch,
    removing all the YAML files and updating the `RELNOTES.md` file
  * Someone reviews and merge the release notes PR

That should save some time when releasing a new version of TPA.
The release notes will be generated an reviewed along with the
changes in the code.

Note: a copy of the `README.md` file was copied from the
release_notes repo to tpa-internal repo because it contains
instructions on how to use the scripts and generate the YAML files
for release notes.

With valuable contributions from Martin.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>