Skip postgres/config/final on replicas during BDRAO upgrade
It should have been like that to begin with.

We filter out 'replica' instances when running postgres/config/final
since the idea around postgres/config/final is to make changes that
involve a running postgres instance, e.g. tasks like creating databases,
extensions, users etc. This is unlike postgres/config that should render
similar settings on both primary and replica so they match. Is there a
scenario where a replica may have some setting that is different from
the primary and could be done on a running postgres instace? Maybe; but
we do not recommend replicas drifting in terms of configuration purely
because a replica is supposed to replace a primary in case of a
failover.

Fixes: TPA-639