Ensure existing edb_repositories are still intact after reconfigure
Make sure that we keep existing edb_repositories entries with
reconfigure command. Exception to this is when `--edb-repositories` is
used, in which case we assume the user will give the complete list (to
allow advanced usage).

Also adds a relnote entry for the changes.

With valuable feedback from Richard and Mario.

References: TPA-660