Improve `relnotes.yml` workflow
.# GitHub token

`edb-slonik` user token (`secrets.GH_SLONIK`) doesn't have enough
rights to create/update PRs in the `tpa-internal` repo.

This commit changes the workflow so it uses `secrets.GITHUBSECURITYTOKEN`
instead of `secrets.GH_SLONIK` when creating/updating the PR.

.# breaking changes type

This commit adds the `breaking_change` relnote type. That type is currently
not used, but might be helpful in the soon future.

.# Add relnote YAML template

We add a relnote template file at `release_notes/relnote.yml.template`. It should
be an easy start when creating release notes files during development.

As a consequence we changed the step `Clean up release notes YAML files` so it only
removes `.yml` and `.yaml` files, so the template won't be removed by the workflow
execution.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>
Co-authored-by: Jonathan Renon <98327055+JonathanRenon-EDB@users.noreply.github.com>