Fix password generation during upgrade
TPA-85 introduced the use-vault script that relies on python `os.getcwd`
to find cluster_dir. `tpaexec upgrade` command defaults to run in
cluster_dir/commands due to ansible default behaviour for playbook. we
need to ensure we are in the correct directory when trying to call
use-vault.

We saw a similar problem in TPA-656 too. There may be other occurences
of this issue for upgrades.

References: TPA-667