Use correct architecture for pulling PGDG repo rpm
The PGDG repo URL for yum contains the architecture string.
It is currently hardcoded to x86_64 as this is the only architecture
TPA supports. This prevents experimentation with (for example) EL
containers on ARM systems - something which works fine with Debian.

This change adds a setup step to detect the architecture and set it as
a fact, then uses it in the URL.