pgd: don't synchronize database structure to witness nodes
In BDR 3.7 ~ 5.x, bdr.join_node_group() has an option
"synchronize_structure", which defaults to "all". This means
that when a witness node is joined to a BDR cluster, the
existing database structure will be copied to the witness node.
However it will never be used or updated. While this is
essentially harmless, it can be a cause of confusion, so
have TPA set "synchronize_structure" to "none" for witness nodes.

TPA-665.