New task_selector scheme to replace Ansible tags
Setting the variables `included_tasks` and `excluded_tasks` to
comma-separated lists of tags will populate the `task_selector` fact,
which tasks can then check with the `selects` and `permits` filters.

The variables can be set in config.yml, on the command line with `-e`,
or on the command line with `--excluded-tasks` and `--included-tasks`,
which are translated to the `-e` equivalents.

The `selects` filter matches if the given tag is explicitly included,
or if there are no explicit inclusions and the tag is not explicitly
excluded. The `permits` filter matches if the tag is not explicitly
excluded.

Fixes: TPA-657