TPA-697: ensure relink manage tower use case
we need to ensure that tpaexec relink does the correct thin when used in an AAP enabled case.

we don't use symlink but hard copy of files such as deploy.yml commands/upgrade.yml

so we need to ensure the cluster has the latest available version of these files (after update of tpaexec version) for instance.

change to relink:
  - check for absence of file
  - compare source and target file bytetobyte
  - ensure we copy in tower scenario
  - still symlink otherwise.