Support cgroupsv2
When we invoke the docker_container module, we now explicitly set
cgroupns_mode to "host".

If the host is running cgroups 1, we mount /sys/fs/cgroup into the
container, marked as ro, as we always have previously. If the host is
running cgroups2, we mount either the whole of /sys/fs/cgroup or one scope
within it into the container, marked as rw. The choice of whether to use
scopes is based on a list of images known to support them: essentially,
RHEL9, Ubuntu 22.04, and images that are based on those.