Update docs and workflows for cgroups2
The docs now give revised instructions for how to use cgroups1 and
explain when it is required.

Workflows are updated to use Ubuntu 22.04, a cgroups2-based image, for
non-RHEL 7 containers when running multiple clusters from one workflow.
Workflows that run a single cluster now default to Ubuntu 22.04 but
allow the user to select Ubuntu 20.04 .