Revert docs/src "style guide compliance" changes en masse
Although these changes represent a substantial amount of work, and it's
a difficult decision to revert them, it's not practical or sustainable
to apply changes to the TPA docs in this fashion. The cost in terms of
errors and increased review load is too high, and the corresponding
benefits far too minor.

There are certainly desirable changes here. Unfortunately, after going
through all of the changes merged to the docs between 23.27 and 23.31,
hoping to keep as much as possible, I concluded that the changes were
organised in a way that makes it impossible to pick them apart with a
reasonable effort.

There are two main problems with the approach:

    1. "Style guide compliance" changes are applied too thickly, even in
       places where they introduce errors rather than fixing them.

    2. Many different kinds of changes are mixed together in the same
       commits, ranging from typo fixes and mechanical changes to many
       unnecessary changes clubbed under "clarity and consistency" that
       are anything but.

In isolation, few of the changes are noteworthy. Many are fine. Many are
only slightly awkward. Most aren't worth arguing about, even when they
introduce some minor ambiguity. Only a few create significant errors.
But taken together, they represent an un-reviewable stream of changes
that the maintainers cannot hope to merge while keeping embarrassing
errors to a minimum (based on already-merged errors, and despite the
fact that many more embarrassing errors were corrected in review).

It's only to be expected that large changes will bring errors with them.
This work should have been done only after discussion about what changes
are important, and how they should be organised to make them easier to
review and merge. This is far more than merely a theoretical concern:
the incidence of errors, large and small, in the work that's *already
been merged* is far too high. The few improvements we want amount to
little more than a handful of minor rewordings.

This is especially unfortunate because there are several places in the
existing docs where the wording is verbose, tedious, and confusing. The
changes reverted here, for the most part, settle for making a few tiny
changes to the minutiae of these unfortunate passages, passing up the
opportunity to make the text better. These changes often bring some
minor grammatical or factual annoyance with them.

Just as we have many good reasons to avoid making wholesale code style
changes to a large existing codebase, and are careful about reducing
churn to avoid conflicts, this exercise conclusively shows that it's
equally important to be cautious with similar changes to docs. The way
to achieve compliance with a style guide, no matter what it says, is to
write new text in the desired style, and change old text when there's a
good independent reason to do so (e.g., to fix errors).

(Yes, this means there will be a long time during which TPA users will
have to see "can not" in some places and "aren't" in others, and suffer
future tense in certain files while being delighted by the clarity of
newer passages that use present tense exclusively.)

It is imperative that future efforts to improve the documentation be
focused on fixing specific problems, clearly stated in advance (and
prioritised in the context of the roadmap, not in isolation). If it's
not worth mentioning in the release notes, it might well not be worth
fixing in the first place.

Furthermore, mechanical changes (such as `-quoting config.yml) should
only ever be applied with a global search-and-replace in a single
commit, to simplify review and management.

This reverts the following commits:

    69b2afba5ad8071a5887c4d2df445fde343bf7d4
    eb8efddd4852aa9fd1c3ada89439529ba325ddee
    bf5a8895dbe1f734fd9b5874eb7b36f744efb5e6
    1d4051b4cdd42ed70c883fc7057c842247929f39
    682805c36d12acd9717296a2e88246072df38471
    01b66c788394a77161d0698884ed1a37f5cd0417
    9ab53455c5ec2b5a866c4cf87f77d7dd4ab93bc8
    f8ce1783333212bc7c07ac7919eadcd75879d73b
    4225edcc5656c4e495f439eb58d6acfa9262d00c
    25eb5a8283b8bca5db6d67ab6de156a1719b17ce
    c0972d96acd1c8cf971c6b19869862ef2e77b6c4
    bf518a0d528c98539d2178ff34b4139521b72ea2
    1b090eae827552c76633ebb1cf48ee73088defad
    93fa459d14b4bb75a26fef872a87bf6356ca0406
    34ed00716684af9abae6c42acb37b8e459a8e2df
    b76fd5f39cb8a235b3d5f4c770e17db1ce10c54f
    a320424d732b5bfb7810e436b0a642759570936d
    3bd11f99fab452494a288b8d10056f6f8d2a828e
    774ea0a65e7dc3db157d16220c3ec935580d3704
    5cac2858c942fb685b7187634731dc61c45642d4
    1182c86a5112eb66ad88c9159c960db98815c4c9
    fc5ecfda0fae0981a199fb95c8b4bf220cd97b26
    89997381e783fb5fae197e9e47078bc0900874bc
    d54e99111fccaf9165a41ef027fa2261657c725a
    c1ce266866bdd514c51b806b2a3a3733d70649b3
    3a195c22c9b8cd05ea9717d2376b73206cbbe9cd
    14dee3bb20dda769a3318c0d65973799fd8e299b
    6aabb0811d5daad425b83d568ed89cefc57a0cee
    5eacf5a8ca175b04ae94dc5452d30d600f67bb29
    7ec58ae2b1658e1d79c092bbecf2a29b9803623c
    ba6f7c0a70af144fb2fa58d8154b70ea47de2fde
    272a17c84a3434e68850e5372341dc59496db0ed
    499abc8ec0b577bf7323b82921ec03bd8903ddd4
    1478879898311b7766f6d02d8f3a2cd1f687892a
    9b54b7abc920f189d7c97a6627b0ad17cfe3f605